% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.query-package.R
\docType{package}
\name{gsm.query-package}
\alias{gsm.query}
\alias{gsm.query-package}
\title{gsm.query: Query Data Plugin}
\description{
Query clinical data modules for use in the `gsm.app` package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OpenRBQM/gsm.query}
  \item \url{https://openrbqm.github.io/gsm.query/}
  \item Report bugs at \url{https://github.com/OpenRBQM/gsm.query/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

Authors:
\itemize{
  \item Jeremy Wildfire \email{jwildfire@gmail.com}
  \item Laura Maxwell \email{laura.maxwell@atorusresearch.com}
  \item Spencer Childress \email{spencer.childress@gilead.com}
}

}
\keyword{internal}
